requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Redis & Caching
redis==4.6.0  # Compatible with celery[redis]
aioredis==2.0.1
celery[redis]==5.3.4

# AI & ML
google-generativeai==0.3.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# OCR
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Notifications
twilio==8.10.3
requests==2.31.0

# PDF Generation
reportlab==4.0.7
WeasyPrint==60.1

# Excel/CSV
openpyxl==3.1.2
xlsxwriter==3.1.9

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Background Tasks
celery[redis]==5.3.4
flower==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26

# Logging & Monitoring
loguru==0.7.2

# Translation/i18n
# googletrans==4.0.0rc1  # Commented out due to dependency conflicts - use alternative
